scripts/copaw/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "copaw"
version = "0.0.7"
description = "CoPaw is a personal assistant that runs in your own environment."
readme = "README.md"
requires-python = ">=3.10, <3.14"
dependencies = [
    "agentscope==1.0.16.dev0",
    "agentscope-runtime==1.1.0",
    "httpx>=0.27.0",
    "discord-py>=2.3",
    "dingtalk-stream>=0.24.3",
    "uvicorn>=0.40.0",
    "apscheduler<4,>=3.11.2",
    "playwright>=1.49.0",
    "questionary>=2.1.1",
    "mss>=9.0.0",
    "reme-ai==0.3.0.6b3",
    "transformers>=4.30.0",
    "python-dotenv>=1.0.0",
    "python-socks>=2.5.3",
    "onnxruntime<1.24",
    "lark-oapi>=1.5.3",
    "python-telegram-bot>=20.0",
    "twilio>=9.10.2",
    "pywebview>=4.0",
    "aiofiles>=24.1.0",
    "paho-mqtt>=2.0.0",
    "matrix-nio>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=4.2.0",
    "pytest-cov>=6.2.1",
]
local = [
    "huggingface_hub>=0.20.0",
]
llamacpp = [
    "copaw[local]",
    "llama-cpp-python>=0.3.0",
]
mlx = [
    "copaw[local]",
    "mlx-lm>=0.10.0; sys_platform == 'darwin'",
]
ollama = [
    "ollama>=0.6.1",
]
full = [
    "copaw[llamacpp,local,ollama]",
    "mlx-lm>=0.10.0; sys_platform == 'darwin'",
]

[project.scripts]
copaw = "copaw.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]
